General Parameters working reasonably well ...
This version works reasonably well as long as no mistakes are made.
Some of the error recovery for things like bad parameter names is
still lacking. But this makes a good basis for testing.

It would be ready to push except that it's not clear yet how the
ParameterSpace code should be integrated. Right now it is a
stand-alone module and that has some advantages (like separate
testing outside of Blender). But it's not clear if that justifies
it being in it's own file or not.