Adding General Parameters (experimental version)
The parameter code is mostly working except I took a few shortcuts on
the error checking. Specifically, the program will not allow you to
create duplicate parameter names. If you enter a name that already
exists, it will be changed back to the previous name instantly.
I tried to create a visual warning, but the discovery of the duplicate
is done in the "update_parameter_name" callback, and I couldn't find a
way to generate a report from within that function. Another shortcut is
that the program will not allow you to delete a parameter that is used
in another expression. Because this is handled by an operator
(rather than a callback), it does flash a warning message in the
notification area to let you know which other parameters are using it.
This isn't too unreasonable. There's also a problem with using certain
Python keywords as parameter names (names like "if" or "and"). I used
Python's built-in parsing and evaluation functions which generate errors
when they see an expression like "if = 5". Finally, if you use a name
that is not defined, the undefined name is replaced with the somewhat
obnoxious value: "(0*1111111111*0)". This value was chosen because it
evaluates to zero (a reasonable default) and it's very easy to see.
Unfortunately this is not done for parameters that are used before
they're defined. They currently create an undefined expression that
is not flagged. Their value will reflect the last valid value.

So the summary for this version is that it should work properly for
properly entered input, but the error checking is much less than desired.