General Parameters Beta Test
The general parameters appear to be functional. They can be created,
modified, renamed, and deleted. They can contain both normal and
"forward" references (although forward references may be disallowed in
future versions).

This version prints out the entire parameter space (to the console) when
any parameter is changed (for debugging). It also displays the parameter
"ID number" for each parameter in the panel (also for debugging).

New parameters are intended to get a default value of zero, but they are
sometimes displaying a value of "None" instead. This is likely a bug,
and I'll be looking into it soon.

Other than that, this code shouldn't break anything (other than
requiring the "ParameterSpace.py" file as part of the addon).