Time Step uses Parameters.
This is the first version of CellBlender to use parameters in actual
fields.

The MCell time step is implemented with 2 properties:

    time_step (FloatProperty)
    time_step_str (StringProperty)

Originally (and eventually) only the string version is shown in the panel.
However, this version also displays the float version for debugging.
Most of these types of parameters call the "check_val_str" function which
checks the string value, restricts it to the current range settings, and
returns the restricted value.

I added a new function similar to "check_val_str" called "check_expr_str"
which takes the same arguments plus an "mcell" argument up front. It attempts
to evaluate the string using the parameter space structure stored in the mcell
object. If it fails, it falls back to the previous evaluation using float(str).
It returns the same value and status as returned by "check_val_str", so the only
change needed to use it is to add the "mcell" argument.

At this time, it works fine with regard to setting the Time Step based on an
expression that involves parameters. However, changing the value of a parameter
used in the Time Step expression DOES NOT currently update the Time Step display.
I believe this has something to do with the way that Blender detects Property
changes. There may be a way to trigger this, but I haven't figured it out yet.
So for now, you must manually trigger the update in the Time Step property to
see the new value. Here's an example:

  1. Create a new parameter called "dt" and set it equal to 0.00001
  2. Set the Time Step to "dt"

With those values, you will see something like: Time Step=9.999999747378e-06.
This shows the actual float value (unrounded), and you may need to enlarge
your window to see the exponent. However, the float version shown below it
should show Blender's rounded version:

  "Experimental View of Time Step: 0.00001"

Unfortunately (for now), if you then change the Parameter setting for "dt" to
some other value, the change is not reflected by the Time Step parameter. To
see the new value, you must click in the Time Step parameter box and then hit
"Enter" to force it to recalculate. I still need to find a way to trigger the
redrawing of non-parameter widgets when a parameter is updated (TBD).

Finally, while this version does display an "Experimental Time Step" it is
otherwise as functional as before, so that's why I chose to push it even though
the new features aren't completed yet.

  CellBlender Source ID = cd99ada9fbd5e705a1ffa452a8ae1223253377c7