This is a partial implementation of using general parameters in panels
This version is still experimenting with time step, but if it works
for time step, then it should be applicable to other panel properties
as well.

The original hope was that if we named panel properties using a known
pattern, then we'd be able to search through all properties to find
those that need to be updated when a general parameter changes. But I
was unable to find a way to traverse through all the properties to
make this work.

As a work-around, this version is experiementing with simply hard-
coding the update functions into the parameter modification code.
So when a general parameter changes, it calls the update functions
for all of the panel parameters.

This approach appears to work for time step in this version. So if
time step is made a function of another parameter (say B+5), then
it will be updated when the parameter "B" is changed. However, this
version does not update the expression ("B+5") to "C+5" if "B" is
renamed to "C". For that to work the panel parameter must also be
associated with a parameter variable in the parameter space. Our
convention will be that such "panel parameters" (like time step)
will be given a special name in the parameter space which is not
an otherwise legal MCell variable name (like $time_step). This way,
we can be sure that users cannot create their own parameter name
that would conflict with one of these internal "panel parameters".

The implement and update these special parameters is not yet
included in this version ... because it hasn't been written yet.