More work on parameters / Reuse of IDs in ParameterSpace.py
In addition to trying to work parameters into Blender, this version of
ParameterSpace.py re-uses parameter ID numbers for any parameters that
had been deleted. The previous version uses a shortcut of always
incrementing the parameter ID, and it was only reset (to 1) when all
parameters were deleted. However, that cannot happen when we begin to
use "panel parameters" because they will always exist (whether any user
defined parameters exist or not).