Merging of Dipak's BioNetGen importing into General Parameters
This version merged Dipak's importer with the general parameters
system. It turns out that this was fairly easy since the general
parameters code was largely derived from Dipak's work.

Note that on my machine, the importing took a very long time due
to the directory search that takes place in "execute_bionetgen".
In most cases, this took over a minute while Blender was locked
up. I ended up hard-coding the directory in this version to my
own path to speed it up. That means this version is NOT suitable
for general use unless this is changed!!!

CellBlender Source ID = b94e3afbecef4648bf8f580c73e5ebd609f7616a