Modified export_mcell_mdl.py to work with parameters
The name of the string versions of parameters was changed from
something_str to PARAM_something to be consistent with our naming
of panel parameters in the user interface: begin with "PARAM_".
Initially, it was thought that these string values weren't used
anywhere other than the user interface, and that the actual
numeric versions were used in other locations. However, the
export_mcell_mdl.py file checks the string versions to see
if they are empty before using the numeric values. Typically:

  if init.something_str:
     out_file.write ("SOMETHING=%g\n" % (init.something))

This was changed (in certain places) to:

  if init.PARAM_something:
     out_file.write ("SOMETHING=%g\n" % (init.something))