Fixed "update_all_panel_parameters" to properly step through molecules
This version contains a lot of debug (and other unneeded) code to try to
find a problem that resided in the "update_all_panel_parameters" function.

The update_all_panel_parameters function is a function that forces the
redrawing of all panel parameters by setting them equal to themselves.

In the case of molecules, this means looping through all of the
molecules in the list and assigning their parameters (such as
diffusion_constant) to themselves. In the case of the diffusion_constant
(for example), this will force a call to "update_diffusion_constant".
However, update_diffusion_constant uses the active_mol_index to decide
which molecule is being changed.

The current "fix" is to set the active_mol_index to each molecule before
calling update_diffusion_constant, and then restoring it to its original
value after the loop. This appears to work.

However, a better solution might be refactoring update_diffusion_constant
so that the actual updating function is separated from the callback
function. This way, both the callback function and the looping function
could call the same shared update function without the need to change
the active_mol_index. But it's not clear if this wouldn't cause other
problems (such as Blender not updating its display properly). So for
now the code will continue to step through the molecules by changing the
active_mol_index.