Working ... but with an error message.
This appears to propagate name changes into the panel parameters,
but it generates an error message when it recognizes that the old
expression is no longer valid (after the name change), so it
regenerates the expression from the IDs (as it should). But this
isn't a very good way to make it work, and it's not clear how to
disable the error message for just that case.