These are from Git Commit 597f4dfcc8b3 before adding panel parameters
The previous implementation of panel parameters attempted to simply
substitute a string expression for the existing parameter (which might
have been a string or a float or an int). That approach was flawed
because using a simple string for an expression was not able to store
the dependencies needed when parameters change names.

The newer approach will be based on the PanelParameter class which
can contain this information as well as member functions for better
encapsulation of the needed functionality. However, this will require
lots of souce code changes which may temporarily break functionality.

So this older version was pulled out as the baseline for a new Git
branch to explore this new way of handling parameters. This same older
version will also be updated with Dipak's recent changes to the
export_mcell_mdl.py file so it can be used as the trunk again while
the "property_parameters" branch undergoes development. When the
new parameter branch is ready, it can be re-merged with the trunk.

CellBlender Source ID = 8b4d93ed5f0e4171c9d6b138f748a03d08c81a30