Backing out Parameter Panel Changes to Git Commit 597f4dfcc8b3 +
The previous implementation of panel parameters attempted to simply
substitute a string expression for the existing parameter (which might
have been a string or a float or an int). That approach was flawed
because using a simple string for an expression was not able to store
the dependencies needed when parameters change names.

The newer approach will be based on the PanelParameter class which
can contain this information as well as member functions for better
encapsulation of the needed functionality. However, this will require
lots of souce code changes which may temporarily break functionality.

So this older version was pulled out as the baseline for a new Git
branch to explore this new way of handling parameters. This version
should have Dipak's most recent changes (f524ea2a21db).

Finally, it was tricky to back out the changes, so please be wary
of this version until you're sure that it is otherwise current with
any changes you've been making. I apologize in advance if I've
clobbered anything inadvertantly.