Simulation runner works with Windows now.
The Python path for the simulation runner was being determined by
shutil.which. This will probably not work in Windows unless it's in the
PATH environment variable (I haven't tested this yet). Regardless, we
shouldn't assume that people are going to edit their PATH. Therefore, I've
added a method to select the location of Python, much like we already do
with MCell.

There was also a problem with run_simulations.py with regard to how
multiprocessing works in Windows. This has been fixed. However, we should
still investigate better long term solutions to some of simulation runners
other problems (e.g. can't kill jobs).