Add support for import of MDL geometry files.  For now such MDL files must contain only valid geometry syntax (nested OBJECT, POLYGON_LIST, BOX).  Geometric transformations and surface region definitions within the geometry blocks are also allowed.  Transformations are correclty applied, and regions are defined. Molecules added to the surface regions are currenlty ignored silently.
The MDL import is implemented in a stripped down version of MCell's MDL parser
in c/flex/bison.  The python interface to this is generated by SWIG and
compiledinto a python module called
"cellblender.io_mesh_mcell_mdl.mdlmesh_parser".  However the core of this
module is a platform-specfic piece called _mdlmesh_parser.so which is generated
by the compiler.  Thus we have a special makefile in io_mesh_mcell_mdl that
handles the build.

To make things slick and easy I have also added a top-level makefile that
descends into io_mesh_mcell_mdl, builds the importer module, and then
packs the whole cellblender addon into a zip file for easy installation.
At the top-level just type "make" and you'll get cellblender.zip.  But if
you also type "make install", the make file will take the zip file and
automagically unpack it into your blender addons directory.  Edit the
top-level makefile to adjust this location.  We could probably figure out
a way to make this more automatic and general.