Found quadratic delay problem - converting lists to be printed
The threshold_print function takes a string, so lists were being
converted to strings before being passed in. This meant that they
were being converted whether they needed to be printed or not.
This current version just commented out the non-essential prints.
The next version will provide a better solution.