Removed the "(new)" from the Parameters Panel.
This version was tested by running through the introduction up
through section 3.6.3. Some of the steps in the introduction are
out of date, but everything worked with minor adaptation to the
documented procedure.

In addition, three general parameters were created - one for the
overall rate (value of 1.0), and one each for volume and surface
molecule (their normal values multiplied by the overall rate).
The simulation was run multiple times with different settings of
the overall rate and the observed diffusion behavior was as
expected. This is not a thorough test, but enough to justify
pushing it for the team to check out.

CellBlender Source ID = 7db135ab36594eeafb6d9ffcec679d175e914eca