Converted iterations to a panel parameter expression
Previous panel parameters had been added by rewriting their
Property/Operator/Panel code to become a single class. This
is important, but somewhat labor intensive. So in order to
provide expression functionality more quickly (with the
upcoming UCSD class in mind), this version simply uses the
panel parameter objects inside the existing structure. This
required the following changes:

  Adding a "set_defaults" method to MCell Property Group
  Adding a "set_defaults" method to MCell Initialization
  Adding a "set_defaults" function to cellblender_operators.py
  Adding an initialization flag to MCell Property Group
  Adding "contains_cellblender_parameters" flags

The set_defaults functions are currently needed because
PointerProperties do not provide a full set of initialization
parameters as provided by Int and Float Properties. This is a
minor drawback to the current approach, and it is hoped that
better initialization mechanisms for PointerProperties will be
available in future versions of Blender.

This version has been minimally tested, so more testing is
welcome.

CellBlender Source ID = 3b487c2f5086453be52ffaefdc42cdfdf2d2e6fc