Parameterized time_step_max
Since this value is intended to only be exported when non-empty,
the expression itself is checked to see if it is empty before
being output to the MDL file by the MDL exporter.

Also note that all current exporting is being done using the
numerical values rather than the expressions ... for now.

CellBlender Source ID = 94be78734743620f5122d7472ab35b1308ee3a6e