Moved .param_data.expression to .expression
This is the first in a series of changes that are needed to support
tool tip descriptions in Blender. The problem with the previous
approach was that the description field in the expression StringProperty
was not easily accessible after creation. Furthermore the description
seemed to be a class variable meaning that changing the description for
one object changed it for all objects. This pretty much requires a new
class for every panel parameter that needs a different description.

Creating a new class for each parameter might be done dynamically, but
at this point, it seems easier to just create a new class for each
panel parameter that needs a new description. Using inheritance, we
can hopefully minimize the amount of code needed for each such class.

This version will probably NOT run, but it is being saved to keep
track of the changes done to the code on a fine-grained basis.