This appears to be a working version with tool tips.
By putting the expression into both the PanelParameterInt and
PanelParameterFloat (in addition to the PanelParameter), it
seems to work with older (non-classed) parameters as well as
the newer (classed) parameters.

This certainly needs more testing, but it means that the
conversion can happen without breaking existing functionality.

Note that this version used an over-ridden version of
PanelParameterFloat during experimentation. This will be
changed next.