Molecules with Tool Tips / Removed "name=" parameter in PointerPropertys
Added Tool Tip functionality to Molecules by adding a class for each
parameter.

Removed the "name=" parameter when defining PointerProperties. The name=
parameter appeared to be redundant, and I couldn't find where it was
being used within Blender. Since that information is already in the new
class for each parameter, it was removed from the PointerProperty
definition. If there was a reason to have it in the PointerProperty
definition, please let me know.

Bob Kuczewski