Set default release probability to 1 and tested on cube demo.
Everything appeared to work, but because I had changed the release
probability to use a parameter, I experimented with different
values for that probability (something I hadn't done before).
Surprisingly, for release probabilities less than 0.9 releasing
2000 vol1 and surf1 molecules, NONE were actually released. I
checked the MDL files to verify that 0.9 was being written to
the file, and it was. So this might be something that either I
don't understand about MCell or something to look into.

I also noticed a delay between pressing "Run" and actually
seeing any change. I checked my machine and it was being heavily
used by MatLab jobs, so that might be the problem.

Other than that, this appears to be working reasonably well...

CellBlender Source ID = 0d8ee888d7a9d170ce19150862753868a4f458cd