Speeding up parameter API and simple MDL importing example
Did some work on trying to speed up parameter creation to support
importing large numbers of parameters (it's still too slow).

This version also contains a toy demonstration on how to read
parameters from an MDL file. This toy demonstration is NOT a
full MDL parser, and it makes lots of assumptions about the
formatting of the MDL file. But it does provide a simple one
file example. It was created in the "mdl" subdirectory to be
parallel with the "bng" subdirectory for BNGL importing. We
might want to create one subdirectory where we put all of
the importers (similar to the "data_plotters" subdirectory)
so they can be handled uniformly ... especially if we're
anticipating more of them.

This version has gotten very little testing, so ... beware!!

CellBlender Source ID = 91b4c1492f260b3444424c4d73bb01bdbf18dbf5