Added "to_python" to build_expr_str and translate_panel_param_ID_expr
Evaluation of MDL expressions in CellBlender is done by converting them
to Python expressions first. This was being done for general parameters
but not for panel parameters. As a result, any MDL expression that was
equivalent to a Python expression worked properly, but any MDL
expressions that differed ("COS" instead of "cos", for example) were not
properly evaluated. The problem was that the same routine used to
convert an internal expression into an editable string was being used to
convert an internal expression into a python executable string for panel
parameters. As a quick fix, the "build_expr_str" function was given a
"to_python" parameter which is defaulted to False. So most conversions
from internal representation to string will still produce an MDL result
as needed for editing. But when a python result is needed, the flag is
passed as True to perform the conversion.

CellBlender Source ID = 51aa4d5de7bc3eed78f7ea709f0519de83106be7