Added print statements during BNGL import
The BNGL importing process is somewhat of a batch process which
actually runs BioNetGen. This is mostly communicated to the user
via the console which shows the progress of the BioNetGen program
itself as it runs.

The final step in the process is the importing of the parameters,
molecules, reactions, etc. Unfortunately, this final step also
takes time (maybe too much time right now), but there's no user
feedback about what's happening during this period. So in keeping
with the BioNetGen console output, I've added print statements to
the importing process to let the user know that *something* is
happeneing while all that data is being loaded.

CellBlender Source ID = 51aa4d5de7bc3eed78f7ea709f0519de83106be7