Added batch mode parameter addition calls.
This is a bandaid approach to speeding up parameter addition.

It provides two functions to the MCellParametersPropertyGroup:

  start_batch_addition()
  finish_batch_addition()

These may be placed around a series of add_parameter_with_values
calls to disable the internal updating of all parameters as each
new parameter is added.

In my brief experiments, this cut the parameter load time in half.
That's not tremendous improvement, but it is a start.