Added MCell mesh importer using pyparsing.
This does not replace the swig based importer but is just another
option, which has the benefit of not requiring the use of a makefile.
Unfortunately, the pyparsing method is considerably slower. If the swig
based importer is found, then it is used. Otherwise, we fall back on the
pyparsing importer. The pyparsing code is an adaptation of Tom's old
importer that he created for Blender 2.49. It is still missing some of
the features which are found in the swig based importer (e.g. scaling,
translate, rotate), but these can be added over time.

The importing of regions works in both importers. All regions are
colored with a red material. Perhaps, this should be maded optional.