Updated some default values.
After discussing this with the group, we agreed that it made sense to
change some of the default values. Both of these changes are small but
hopefully make CellBlender a little more accessible for new users and
allow for quicker prototyping.

First, there are very few situations when a user actually wants to run a
simulation for zero iterations, so the default is being changed to one.
Since the user won't see any viz data when iterations is set to zero, it
can potentially result in the user hunting around for a bug that isn't
actually there, since it is technically valid input.

Second, we changed it so that errors are printed to the console by
default instead of to a file. This was actually the original setting,
but we changed it since some people do not have a console if they start
Blender by, for instance, double clicking an icon. In practice, however,
it's much easier for the user to simply enable the console than to go
hunting around through a directory tree trying to find an error log.
Ultimately, a better long term solution would be to actually show the
error messages within Blender itself.