Added ability to use variable rate constants.
When creating a reaction, a new check box has been added that allows the
user to select whether they want the reaction to use a variable rate
constant or not. If selected, the user can navigate to an appropriate
text file (time in first column, rate constant in second). The file is
read in and stored as a Blender text object. At export time, a new text
file will be created from the text object and stored in the CellBlender
project directory. Some basic error handling has been added, but it
could be more robust.