Moved the binaries to CB preferences.
Moved the selection of the MCell, BNG, Python, and SBML2MCELL selectors
to "CellBlender - Preferences". Two of the options already under CB
preferences were added to the list of presets that are saved when a new
preset is created. I haven't yet found a way to automatically load the
presets when a blend file is loaded.