Changed how preferences/presets are saved/loaded.
Preferences are now saved automatically as they are changed. There is
only one set of preferences, which is automatically loaded when blender
is started or a new blend file is opened. Since there is only one set,
which is saved and loaded automatically, the drop down menu and
accompanying +/- buttons were removed. A button was added which lets the
user reset the preferences to their default values.