Added button to remove all regions on an object.
This commit resolves issue 16:
https://code.google.com/p/cellblender/issues/detail?id=16

The button is an X below the standard + and - buttons. The X icon is
used in other instances to indicate deleting all entries. For instance,
both Vertex Groups and Shape Keys use it in this manner although in
those cases the X icon is found in a drop down menu beneath the +/-
buttons (along with various other related functions). If additional
operations are required in the future, then it might make sense to
implement them similarly.