Add error msg if reaction directionality is wrong.
If reaction directionality is not specified when it is required (i.e. if
there are any surface molecules or a surface class in the reaction), an
error message is shown. Likewise, an error message is also generated if
reaction directionality is specified when it should not be used (i.e.
all volume molecules).