Bugfix: Import over objects with identical name.
This resolves issue 48 in the bug tracker:

"CellBlender will not import over existing objects with the same name.
It will leave the existing object there without any warning/error. If
you try to delete the existing object first, it will bring that object
back when you try to import the new object."