Improved parameter system to fix unresolved references on name changes.
Expressions may contain unresolved parameters. For example, if there is
no "z" parameter, the expression "5+z" would not be able to resolve "z".

In the older system, a name change triggered a re-evaluation of all
parameters which always resolved the problem. However, since the new
system only updates on dependencies, this presents a problem since there
will be no "linkage" between a newly created "z" and the unresolved
expression "5+z" which already exists. That would mean searching the
entire parameter space for every name change or new parameter.

Fortunately, there already was a "param_error_list" which kept track of
any parameters that contained errors. So the easiest thing to do is to
simply "touch" the expression of each parameter in the param_error_list.
This will cause them to re-evaluate themselves. If the name change or
newly added parameter fixes the unresolved reference, then it will be
redrawn with the correct value.