Added the cellblender_source_sha1.txt to git repository.
I think this will automatically "tag" each release by placing
the computed SHA1 for each release in the git repository for
the version that was checked in. I had been doing this by
hand, but it might be better to have the computed version
automatically included so we can more reliably determine
which source files map to an installed version. This assumes
that Blender and CellBlender have been run to compute the
source SHA1 before the commit is done. Otherwise, the commit
might be including the SHA1 from a previous version. But we
should pretty much always run something before committing it
anyway.  : )