Finished parameterizing all but partitions and warning levels.
Also implemented exporting parameters as values or expressions.
The expression exporting is largely untested and is listed as
"experimental" in the CellBlender interface.

The logic for including the MDL viz block in the main or in a
modular MDL file wasn't behaving the same as the other blocks
(possibly because of the export all feature). I changed the
logic slightly, so it might be worth checking to be sure I
didn't change any desired functionality.

CellBlender Source ID = 88a6ceeab55a713fbc35e4cfcbb6eaa703f7ba6c