Added a new option for handling invalid models.
Previously, a user could only filter out invalid/illegal input or simply
ignore it. These options both serve a purpose but are likely not the
desired default behavior. A new option (available in
Preferences->Invalid Policy), which prevents the simulation from
exporting/running if anything is invalid in the model and furthermore
displays all such entries in a list in the "Run Simulation" panel.

This commit makes a considerable number of changes to
export_mcell_mdl.py with the intention of reducing inconsistencies,
boilerplate code, and unnecessary nesting. Although testing was done,
given the number of changes made, it's possible that some bugs were
introduced. Also, prior to this commit, there were errors not being
caught/handled, and this commit does not address those problems (e.g.
releasing into an object in blender but not in the model object list).