Some refinements of Parameter System and re-included Profiling.
CellBlender has been exhibiting some odd timing behavior with regard
to systems with large numbers of parameters, molecules, and reactions.
For example, with fceri_ji_comp.bngl, it creates about 8450 parameters
(mostly attached to molecules and reactions). These load fairly quickly,
but having that many parameters appears to slow seemingly unrelated
functionality. For example, with that many parameters loaded you can
create an icosphere and move it around with no noticable slow down.
However, if you decide to change the number of subdivisions on the
newly created icosphere, Blender locks up for well over a minute!!
More interestingly, with profiling added to the parameter code, none
of that time is showing up in calls to the parameter functions. Why
not? What is going on? Are we running up against a property count
limitation? The current parameter system uses about 21 Blender
"Properties" for each parameter (12 strings, 8 booleans, and one
float). Multiplied by 8450, that's over 177,000 Blender Properties.
Could that be a problem in itself? Also, we might need to start
profiling some of the non-parameter code as well.