Option to auto add imported objs to model objects.
When you import an MDL containing meshes, you now have the option to
automatically add all imported objects to the Model Objects list.
Presumably, this would be the option most users would want, so I've made
it the default, although we can change it if anyone feels strongly about
it.