Better error handling for release site objects.
Previously, we only checked to see if an object was in the scene to
determine if it was a valid object to release on/in. With this commit,
we now check whether the object is in the Model Objects list (and also
that the object doesn't have an error associated with it). We may also
want to add more detailed error messages in the future.