Improved error checking for release sites.
This commit fixes a problem in which a release site with a "shape-based"
release is marked invalid/illegal even though it is legal. Specifically,
this happens when a release site is created referencing geometry that
hasn't yet been added to the Model Objects list. This problem was fixed
by checking the release sites whenever model_objects_update is called
(mainly whenever an object is added to the model objects list).

What follows describes the background and motivation for the fix:

Typically, a user only creates a release site that references a mesh
object if that mesh object has already been created and added to the
Model Objects list. However, it's entirely possible that a user will
reverse these operations. It's potentially more likely that a user will
do this when importing reaction networks, release sites, and geometry.
And, given that a large number of release sites may be imported, it
seems all the more important to make this check.