Temporary fix for implementing reaction output sums
Cellblender doesn't currently implement complex reaction output statements of the form

{ 2*COUNT[s5,WORLD] + 2*COUNT[s6,WORLD] ....

However ther implementation is necessary for larger models like the ones imported
from BNGL files. As a temporary solutin I implemented a list called 'complex_rxn_output_list' inside complex_rxn_output_list in cellblender properties
that stores this information. This list is later checked in the export_mdl function
and a rxn_output section with this information is created if necessary. No changes
to the GUI were made. The changes are marked with the #JJT tag for easy removal
once we have a more permanent solution