Changed default molecule type to "Volume Molecule"
The default molecule type was originally set to a surface molecule, but,
after discussions with Tom and Bob, we agreed that it likely makes more
sense for it to be set to a volume molecule. Setting the default to a
volume molecule is much simpler for new users, since they don't even
need any geometry (with Cubic or Spherical [Shell] release sites),
unlike surface molecules.