Bugfix: End seed could be less than start seed
If the end seed is ever set to a value less than the start seed, it is
reset to the the value of the start seed. If the start seed is ever set
to a value greater than the end seed, it is reset to the value of the
end seed.

This resolves the following issue:

https://code.google.com/p/cellblender/issues/detail?id=55