Bugfix: Molecs not shown under some circumstances
If a molecule had already been added to the top level "molecules" object
and it hadn't appeared in the previous frame (e.g. 0), then it could
disappear from the current frame (e.g. 1). It would appear in the
subsequent frame (e.g. 2). Or if you stepped back a frame (e.g. 1), then
you could see it in this original frame.

To the best of my knowledge, here is why this was happening: The
molecule wasn't being added to the mol_viz_list on frame 0 (since it
didn't exist then). Therefore, it wasn't being acted upon by
mol_viz_clear when we stepped to frame 1, which means that it wasn't
being unlinked and relinked. However, it was added to mol_viz_list on
frame 1, so it was cleared by mol_viz_clear whether the next frame was 0
or 2, and as such would appear in those frames.

The solution was to unlink and relink the object on the current frame
regardless of whether it was in the previous frame. It seems to be
resolved with no noticeable performance impact.

I believe this resolves the following issue:

https://code.google.com/p/cellblender/issues/detail?id=23