refactoring. changed the way external models are imported to conform to a common interface
Changed the way files are imported in CellBlender. All importing code should create
a json-file with molecule,parameter,reaction,release and output information.
This file is read in the same way and passed to the appropiate handlers.
bng_operators and sbml_operators now only contain code specific to the creation of
this JSON file.