Added a call to init_properties in the mcell-level build_properties...
Newly imported parameters from the data model were not properly
triggering updates of existing panel parameters. Since importing
a data model is considered to be a whole new model (and not yet
a way of merging models), the top-level call to init_properties
is an easy way to start with a fresh (empty) mcell model.