Variable Rate File added to data model as string: "variable_rate_text"
The variable rate constant file is now stored in the data model as a
single string. This was done to be consistent with the storage of the
text in the .blend file. The new key "variable_rate_text" was added to
the data model for reactions. It will be an empty string when there is
no variable rate constant file in use.

Note that this file could optionally be parsed into a list of time and
rate constant pairs and stored as a list of tuples or a list of lists.
This was NOT done in this revision because it's not clear if the
variable rate file will ALWAYS be in the form of two columns of float
values. Since the text file is much more expressive than two columns of
floats, the entire text file was stored in the data model. This might be
changed prior to the next major release.

As noted in the previous commit message, some of the code that was
previously in cellblender_operators.py and export_mcell_mdl.py has been
moved into cellblender_properties.py and placed inside the
MCellReactionProperty class as part of the ongoing "objectification" of
the code. At some point, the reaction code may be moved into its own
file and will consolidate code from the properties, operators, and
panels files.

CellBlender Source ID = 702de468da0ce0b1f5b206fff35e3d9fe53e3d88