Fixed hash computation for Python2 and Python3
The previous version used the "r" flag for reading the source files to
be hashed. The result of the read was converted to a string using the
"utf-8" option. However, that process produces a different result when
the same code is run under Python2 and Python3. This difference became
an issue when the cellblender_id.py file was being created by the
makefile (where "python" can sometimes refer to Python 2) and then being
checked within Blender which uses Python 3.

The solution was to read the files in "rb" mode and not do any "utf-8"
conversion when computing the hash values.